kdb 1/server/requirements.txt
flask==3.0.3
flask-cors==4.0.1
mysql-connector-python==9.0.0
bcrypt==4.1.3
python-dotenv==1.0.1
